# [3.11] Cache construction of middleware handlers